Fix download and verify (#27)
* add test case for checksum download
* set correct responseType for shasum downloads

The default axios responseType is 'json'. In this case we're downloading
a text file (list of shasums) and a binary file (the signature), so
different responeTypes are required.

* add test for downloading and validating a release
* fix incomplete download errors

The download method now waits until all data has been downloaded
and written to the destination file. Before this, the method was resolved
too early, and as a result, subsequent operations ran on an incomplete
download. This caused operations like verify to fail, because of a
different shasum.

* add gitattributes file to fix checksum test on windows
* add type for shared release variable in tests